Bench Generator testing, old file deleted
